https://wavesplatform.atlassian.net/browse/NODE-1197 move matcher node config to separate object
