+++ it/src/test/scala/com/wavesplatform/it/sync/matcher/CancelOrderTestSuite.scala
@@ -19,10 +19,9 @@ import org.scalatest.{BeforeAndAfterAll, CancelAfterFailure, FreeSpec, Matchers}
 import scala.concurrent.duration._
 import scala.math.BigDecimal.RoundingMode
 import scala.util.{Random, Try}
+import com.wavesplatform.it.sync.matcher.config.MatcherPriceAssetConfig._
 
 class CancelOrderTestSuite extends FreeSpec with Matchers with BeforeAndAfterAll with CancelAfterFailure with NodesFromDocker with ReportingTestName {
-
-  import CancelOrderTestSuite._
 
   override protected def nodeConfigs: Seq[Config] = Configs
 
@@ -92,102 +91,4 @@ class CancelOrderTestSuite extends FreeSpec with Matchers with BeforeAndAfterAll
       (BigInt(matchAmount) * matchPrice / Order.PriceConstant).bigInteger.longValueExact()
     }
 
-}
-
-object CancelOrderTestSuite {
-
-  import ConfigFactory._
-  import com.wavesplatform.it.NodeConfigs._
-
-  private val ForbiddenAssetId = "FdbnAsset"
-  private val Decimals: Byte   = 2
-
-  private val minerDisabled = parseString("waves.miner.enable = no")
-  private val matcherConfig = parseString(s"""
-       |waves.matcher {
-       |  enable = yes
-       |  account = 3HmFkAoQRs4Y3PE2uR6ohN7wS4VqPBGKv7k
-       |  bind-address = "0.0.0.0"
-       |  order-match-tx-fee = 300000
-       |  blacklisted-assets = ["$ForbiddenAssetId"]
-       |  balance-watching.enable = yes
-       |}""".stripMargin).withFallback(minerDisabled)
-
-  private val _Configs: Seq[Config] = (Default.last +: Random.shuffle(Default.init).take(3))
-    .zip(Seq(matcherConfig, minerDisabled, minerDisabled, empty()))
-    .map { case (n, o) => o.withFallback(n) }
-
-  private val aliceSeed = _Configs(1).getString("account-seed")
-  private val bobSeed   = _Configs(2).getString("account-seed")
-  private val alicePk   = PrivateKeyAccount.fromSeed(aliceSeed).right.get
-  private val bobPk     = PrivateKeyAccount.fromSeed(bobSeed).right.get
-
-  val IssueUsdTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = alicePk,
-      name = "USD-X".getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = Decimals,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val IssueWctTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = bobPk,
-      name = "WCT-X".getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = Decimals,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val UsdId: AssetId = IssueUsdTx.id()
-  val WctId          = IssueWctTx.id()
-
-  val wctUsdPair = AssetPair(
-    amountAsset = Some(WctId),
-    priceAsset = Some(UsdId)
-  )
-
-  val wctWavesPair = AssetPair(
-    amountAsset = Some(WctId),
-    priceAsset = None
-  )
-
-  val wavesUsdPair = AssetPair(
-    amountAsset = None,
-    priceAsset = Some(UsdId)
-  )
-
-  private val updatedMatcherConfig = parseString(s"""
-       |waves.matcher {
-       |  price-assets = [ "$UsdId", "WAVES"]
-       |}
-     """.stripMargin)
-
-  private val Configs = _Configs.map(updatedMatcherConfig.withFallback(_))
-
-  def createSignedIssueRequest(tx: IssueTransactionV1): SignedIssueV1Request = {
-    import tx._
-    SignedIssueV1Request(
-      Base58.encode(tx.sender.publicKey),
-      new String(name),
-      new String(description),
-      quantity,
-      decimals,
-      reissuable,
-      fee,
-      timestamp,
-      signature.base58
-    )
-  }
 }
