+++ it/src/test/scala/com/wavesplatform/it/sync/matcher/RoundingIssuesTestSuite.scala
@@ -1,23 +1,18 @@
 package com.wavesplatform.it.sync.matcher
 
-import com.typesafe.config.{Config, ConfigFactory}
-import com.wavesplatform.account.PrivateKeyAccount
-import com.wavesplatform.api.http.assets.SignedIssueV1Request
+import com.typesafe.config.Config
 import com.wavesplatform.it.ReportingTestName
 import com.wavesplatform.it.api.LevelResponse
 import com.wavesplatform.it.api.SyncHttpApi._
 import com.wavesplatform.it.api.SyncMatcherHttpApi._
-import com.wavesplatform.it.sync.CustomFeeTransactionSuite.defaultAssetQuantity
+import com.wavesplatform.it.sync._
+import com.wavesplatform.it.sync.matcher.config.MatcherPriceAssetConfig._
 import com.wavesplatform.it.transactions.NodesFromDocker
-import com.wavesplatform.it.util._
-import com.wavesplatform.transaction.AssetId
-import com.wavesplatform.transaction.assets.IssueTransactionV1
+import com.wavesplatform.state.ByteStr
 import com.wavesplatform.transaction.assets.exchange.{AssetPair, OrderType}
-import com.wavesplatform.utils.Base58
 import org.scalatest.{BeforeAndAfterAll, CancelAfterFailure, FreeSpec, Matchers}
 
 import scala.concurrent.duration._
-import scala.util.Random
 
 class RoundingIssuesTestSuite
     extends FreeSpec
@@ -27,8 +22,6 @@ class RoundingIssuesTestSuite
     with NodesFromDocker
     with ReportingTestName {
 
-  import RoundingIssuesTestSuite._
-
   override protected def nodeConfigs: Seq[Config] = Configs
 
   private def matcherNode = nodes.head
@@ -37,7 +30,13 @@ class RoundingIssuesTestSuite
 
   private def bobNode = nodes(2)
 
-  Seq(IssueUsdTx, IssueEthTx, IssueBtcTx).map(createSignedIssueRequest).foreach(matcherNode.signedIssue)
+  val ethId = bobNode.issue(bobNode.address, "xETH", "x-ETH", someAssetAmount, 8, false, issueFee).id
+
+  val ethBtcPair = AssetPair(
+    amountAsset = Some(ByteStr(s"$ethId".getBytes)),
+    priceAsset = Some(BtcId)
+  )
+  Seq(IssueUsdTx, IssueBtcTx).map(createSignedIssueRequest).foreach(matcherNode.signedIssue)
   nodes.waitForHeightArise()
 
   "should correctly fill an order with small amount" in {
@@ -113,115 +112,4 @@ class RoundingIssuesTestSuite
     }
   }
 
-}
-
-object RoundingIssuesTestSuite {
-
-  import ConfigFactory._
-  import com.wavesplatform.it.NodeConfigs._
-
-  private val ForbiddenAssetId = "FdbnAsset"
-  private val Decimals: Byte   = 2
-
-  private val minerDisabled = parseString("waves.miner.enable = no")
-  private val matcherConfig = parseString(s"""
-                                             |waves.matcher {
-                                             |  enable = yes
-                                             |  account = 3HmFkAoQRs4Y3PE2uR6ohN7wS4VqPBGKv7k
-                                             |  bind-address = "0.0.0.0"
-                                             |  order-match-tx-fee = 300000
-                                             |  blacklisted-assets = ["$ForbiddenAssetId"]
-                                             |  balance-watching.enable = yes
-                                             |}""".stripMargin).withFallback(minerDisabled)
-
-  private val _Configs: Seq[Config] = (Default.last +: Random.shuffle(Default.init).take(3))
-    .zip(Seq(matcherConfig, minerDisabled, minerDisabled, empty()))
-    .map { case (n, o) => o.withFallback(n) }
-
-  private val aliceSeed = _Configs(1).getString("account-seed")
-  private val alicePk   = PrivateKeyAccount.fromSeed(aliceSeed).right.get
-
-  private val bobSeed = _Configs(2).getString("account-seed")
-  private val bobPk   = PrivateKeyAccount.fromSeed(bobSeed).right.get
-
-  val IssueUsdTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = alicePk,
-      name = "USD-X".getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = Decimals,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val UsdId: AssetId = IssueUsdTx.id()
-
-  val IssueEthTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = bobPk,
-      name = "ETH-X".getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = 8,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val EthId: AssetId = IssueEthTx.id()
-
-  val IssueBtcTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = alicePk,
-      name = "BTC-X".getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = 8,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val BtcId: AssetId = IssueBtcTx.id()
-
-  val wavesUsdPair = AssetPair(
-    amountAsset = None,
-    priceAsset = Some(UsdId)
-  )
-
-  val ethBtcPair = AssetPair(
-    amountAsset = Some(EthId),
-    priceAsset = Some(BtcId)
-  )
-
-  private val updatedMatcherConfig = parseString(s"""
-                                                    |waves.matcher {
-                                                    |  price-assets = ["$UsdId", "$BtcId", "WAVES"]
-                                                    |}
-     """.stripMargin)
-
-  private val Configs = _Configs.map(updatedMatcherConfig.withFallback(_))
-
-  def createSignedIssueRequest(tx: IssueTransactionV1): SignedIssueV1Request = {
-    import tx._
-    SignedIssueV1Request(
-      Base58.encode(tx.sender.publicKey),
-      new String(name),
-      new String(description),
-      quantity,
-      decimals,
-      reissuable,
-      fee,
-      timestamp,
-      signature.base58
-    )
-  }
 }
