+++ it/src/test/scala/com/wavesplatform/it/sync/matcher/TradeBalanceAndRoundingTestSuite.scala
@@ -1,27 +1,23 @@
 package com.wavesplatform.it.sync.matcher
 
-import com.typesafe.config.{Config, ConfigFactory}
-import com.wavesplatform.account.PrivateKeyAccount
-import com.wavesplatform.api.http.assets.SignedIssueV1Request
+import com.typesafe.config.Config
 import com.wavesplatform.it.ReportingTestName
 import com.wavesplatform.it.api.AssetDecimalsInfo
 import com.wavesplatform.it.api.SyncHttpApi._
 import com.wavesplatform.it.api.SyncMatcherHttpApi._
 import com.wavesplatform.it.sync.CustomFeeTransactionSuite.defaultAssetQuantity
 import com.wavesplatform.it.sync._
+import com.wavesplatform.it.sync.matcher.config.MatcherPriceAssetConfig._
 import com.wavesplatform.it.transactions.NodesFromDocker
 import com.wavesplatform.it.util._
 import com.wavesplatform.matcher.model.LimitOrder
-import com.wavesplatform.transaction.AssetId
-import com.wavesplatform.transaction.assets.IssueTransactionV1
+import com.wavesplatform.state.ByteStr
 import com.wavesplatform.transaction.assets.exchange.OrderType.{BUY, SELL}
 import com.wavesplatform.transaction.assets.exchange.{AssetPair, Order, OrderType}
-import com.wavesplatform.utils.Base58
 import org.scalatest.{BeforeAndAfterAll, CancelAfterFailure, FreeSpec, Matchers}
 
 import scala.concurrent.duration._
 import scala.math.BigDecimal.RoundingMode
-import scala.util.Random
 
 class TradeBalanceAndRoundingTestSuite
     extends FreeSpec
@@ -31,8 +27,6 @@ class TradeBalanceAndRoundingTestSuite
     with NodesFromDocker
     with ReportingTestName {
 
-  import TradeBalanceAndRoundingTestSuite._
-
   override protected def nodeConfigs: Seq[Config] = Configs
 
   private def matcherNode = nodes.head
@@ -41,7 +35,13 @@ class TradeBalanceAndRoundingTestSuite
 
   private def bobNode = nodes(2)
 
-  Seq(IssueUsdTx, IssueWctTx, IssueEthTx).map(createSignedIssueRequest).foreach(matcherNode.signedIssue)
+  val ethId = bobNode.issue(bobNode.address, "xETH", "x-ETH", someAssetAmount, 8, false, issueFee).id
+
+  val ethWavesPair = AssetPair(
+    amountAsset = Some(ByteStr(s"$ethId".getBytes)),
+    priceAsset = None
+  )
+  Seq(IssueUsdTx, IssueWctTx).map(createSignedIssueRequest).foreach(matcherNode.signedIssue)
   nodes.waitForHeightArise()
 
   "Alice and Bob trade WAVES-USD" - {
@@ -340,122 +340,4 @@ class TradeBalanceAndRoundingTestSuite
       (BigInt(matchAmount) * matchPrice / Order.PriceConstant).bigInteger.longValueExact()
     }
 
-}
-
-object TradeBalanceAndRoundingTestSuite {
-
-  import ConfigFactory._
-  import com.wavesplatform.it.NodeConfigs._
-
-  private val ForbiddenAssetId = "FdbnAsset"
-  val Decimals: Byte           = 2
-
-  private val minerDisabled = parseString("waves.miner.enable = no")
-  private val matcherConfig = parseString(s"""
-                                             |waves.matcher {
-                                             |  enable = yes
-                                             |  account = 3HmFkAoQRs4Y3PE2uR6ohN7wS4VqPBGKv7k
-                                             |  bind-address = "0.0.0.0"
-                                             |  order-match-tx-fee = 300000
-                                             |  blacklisted-assets = ["$ForbiddenAssetId"]
-                                             |  balance-watching.enable = yes
-                                             |}""".stripMargin).withFallback(minerDisabled)
-
-  private val _Configs: Seq[Config] = (Default.last +: Random.shuffle(Default.init).take(3))
-    .zip(Seq(matcherConfig, minerDisabled, minerDisabled, empty()))
-    .map { case (n, o) => o.withFallback(n) }
-
-  private val aliceSeed = _Configs(1).getString("account-seed")
-  private val bobSeed   = _Configs(2).getString("account-seed")
-  private val alicePk   = PrivateKeyAccount.fromSeed(aliceSeed).right.get
-  private val bobPk     = PrivateKeyAccount.fromSeed(bobSeed).right.get
-
-  val usdAssetName = "USD-X"
-  val wctAssetName = "WCT-X"
-  val ethAssetName = "ETH-X"
-
-  val IssueUsdTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = alicePk,
-      name = usdAssetName.getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = Decimals,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val IssueWctTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = bobPk,
-      name = wctAssetName.getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = Decimals,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val IssueEthTx: IssueTransactionV1 = IssueTransactionV1
-    .selfSigned(
-      sender = alicePk,
-      name = ethAssetName.getBytes(),
-      description = "asset description".getBytes(),
-      quantity = defaultAssetQuantity,
-      decimals = 8,
-      reissuable = false,
-      fee = 1.waves,
-      timestamp = System.currentTimeMillis()
-    )
-    .right
-    .get
-
-  val UsdId: AssetId = IssueUsdTx.id()
-  val WctId: AssetId = IssueWctTx.id()
-  val EthId: AssetId = IssueEthTx.id()
-
-  val wctUsdPair = AssetPair(
-    amountAsset = Some(WctId),
-    priceAsset = Some(UsdId)
-  )
-
-  val wctWavesPair = AssetPair(
-    amountAsset = Some(WctId),
-    priceAsset = None
-  )
-
-  val ethWavesPair = AssetPair(
-    amountAsset = Some(EthId),
-    priceAsset = None
-  )
-
-  val wavesUsdPair = AssetPair(
-    amountAsset = None,
-    priceAsset = Some(UsdId)
-  )
-
-  private val updatedMatcherConfig = parseString(s"""waves.matcher.price-assets = ["$UsdId", "WAVES"]""".stripMargin)
-
-  private val Configs = _Configs.map(updatedMatcherConfig.withFallback(_))
-
-  def createSignedIssueRequest(tx: IssueTransactionV1): SignedIssueV1Request = {
-    import tx._
-    SignedIssueV1Request(
-      Base58.encode(tx.sender.publicKey),
-      new String(name),
-      new String(description),
-      quantity,
-      decimals,
-      reissuable,
-      fee,
-      timestamp,
-      signature.base58
-    )
-  }
 }
